+++ example/example.dart
@@ -4,6 +4,7 @@ import 'package:pexels_photos_videos/pexels_media.dart';
 import 'package:pexels_photos_videos/photo.dart';
 import 'package:pexels_photos_videos/search_photo.dart';
 import 'package:pexels_photos_videos/video.dart';
+import 'package:pexels_photos_videos/search_video.dart';
 
 void main() async {
   /// Call [PexelsMedia] and get [PexelsResult]
@@ -12,11 +13,13 @@ void main() async {
   Photo photo = await pexelsMedia.getPhoto('image Id');
   Video video = await pexelsMedia.getVideo('video Id');
   Curated curated = await pexelsMedia.getCuratedPhotos(1, 1);
-  SearchPhoto searchResult = await pexelsMedia.searchPhotos("search text");
+  SearchPhoto searchPhotos = await pexelsMedia.searchPhotos("search text");
+  SearchVideo searchVideos = await pexelsMedia.searchVideos("nature", 1, 10);
   if (kDebugMode) {
     print(photo.photographer);
     print(video.user);
     print(curated.totalResult);
-    print(searchResult.totalResults);
+    print(searchPhotos.totalResults);
+    print(searchVideos.totalResults);
   }
 }